refactor, re-run for temporal regions

+++ scripts/multi_session_analysis/20250208_generate_all_units_for_sessions.py
@@ -8,6 +8,7 @@ import utils.behavioral_utils as behavioral_utils
 import utils.spike_utils as spike_utils
 import os
 import argparse
+from distutils.util import strtobool
 
 SESSIONS_PATH = "/data/patrick_res/sessions/{sub}/valid_sessions.pickle"
 OUTPUT_PATH = "/data/patrick_res/firing_rates/{sub}/all_units.pickle"
@@ -17,8 +18,7 @@ def get_units_for_session(row, args):
     sess_units = spike_general.list_session_units(None, args.subject, sess_name, species_dir="/data")
     sess_units["session"] = sess_name
     sess_units["PseudoUnitID"] = int(row.session_name) * 100 + sess_units.UnitID
-    if len(sess_units) > 0:
-        print(f"session {sess_name}: {sess_units.UnitID.nunique()} units")
+    print(f"session {sess_name}: {sess_units.UnitID.nunique()} units")
 
     return sess_units
 
@@ -31,14 +31,20 @@ def generate_all_units(args):
         all_units = pd.concat(sessions.apply(lambda x: get_units_for_session(x, args), axis=1).values)
     print(f"{len(all_units)} units total")
     print(f"across {all_units.session.nunique()} sessions")
-    all_units.to_pickle(args.output_path)
+    if not args.dry_run: 
+        print(f"saving to {args.output_path}")
+        all_units.to_pickle(args.output_path)
 
 def main():
     parser = argparse.ArgumentParser()
     parser.add_argument('--subject', default="SA", type=str)
     parser.add_argument('--sessions_path', default=SESSIONS_PATH, type=str)
     parser.add_argument('--output_path', default=OUTPUT_PATH, type=str )
+    parser.add_argument('--dry_run', default=True, type=lambda x: bool(strtobool(x)))
+
     args = parser.parse_args()
+
+    print(f"Running in dry run {args.dry_run}")
     generate_all_units(args)
 
 if __name__ == "__main__":
